Rename service, rubocop
